+++ lib/src/extensions/numbers.dart
@@ -4,38 +4,6 @@ import 'dart:math';
 import 'package:dart_helper_utils/src/exceptions/exceptions.dart';
 import 'package:dart_helper_utils/src/src.dart';
 
-/* SUGGESTIONS
-1. **Status Code Ranges:**
-   - Methods to check for various HTTP status code ranges, such as informational responses, redirection messages, client errors, and server errors. For example, methods like `isInformational`, `isSuccess`, `isRedirection`, `isClientError`, and `isServerError`.
-
-2. **Specific Status Checks:**
-   - Methods for commonly checked status codes, such as `isBadRequest`, `isUnauthorized`, `isForbidden`, `isNotFound`, `isMethodNotAllowed`, `isNotAcceptable`, `isRequestTimeout`, `isConflict`, `isGone`, `isInternalServerError`, `isNotImplemented`, `isBadGateway`, `isServiceUnavailable`, and `isGatewayTimeout`.
-
-3. **Category-Based Extensions:**
-   - You could also categorize the status codes by their first digit and create properties like `is1xx`, `is2xx`, `is3xx`, `is4xx`, and `is5xx` for a quick range check.
-
-4. **Custom Status Code Creators:**
-   - Functions that allow users to create custom status codes that might not be part of the standard HTTP statuses, which can be useful for APIs that use custom codes.
-
-5. **Retriable Status Check:**
-   - A method like `isRetriable` that returns true for status codes where a retry might be appropriate, such as network-related errors or server errors (503 Service Unavailable, 504 Gateway Timeout).
-
-6. **Redirection Helpers:**
-   - Additional helpers for common redirection status codes, like `isPermanentRedirect` and `isTemporaryRedirect`.
-
-7. **More Informative Enums:**
-   - Expanding the `HttpResStatus` enum to include more descriptive fields for each status, such as a human-readable message or description. For example, `success` could have a description like "The request has succeeded."
-
-8. **Conversion to Exception:**
-   - A method that converts a non-successful status code to an appropriate exception, which can be thrown in the application to handle HTTP errors more effectively.
-
-9. **Deprecation Notice:**
-   - If your API or application plans to deprecate certain endpoints, a method to mark status codes as deprecated could also be helpful, prompting developers to take notice and potentially update their code.
-
-10. **Retry-After Duration:**
-    - If the number represents a `Retry-After` status, you could include a method to parse this and provide a `Duration` object, suggesting after how much time one should retry the request.
-*/
-// TODO(ANY): add more http helpers from response code.
 extension DHUHttpEx on num? {
   bool get isSuccessHttpResCode => this == 200 || this == 201;
 
@@ -71,6 +39,25 @@ extension DHUNullSafeNumExtensions on num? {
 
   /// Returns `true` if this integer is greater than *0*.
   bool get asBool => (this ?? 0) > 0;
+
+  /// Returns a string representation of the number with a given number of
+  /// decimal places.
+  ///
+  /// If [keepTrailingZeros] is `true`, trailing zeros will be kept in the
+  /// result, otherwise, they will be removed along with any unnecessary
+  /// decimal points.
+  String toDecimalString(int decimalPlaces, {bool keepTrailingZeros = false}) {
+    var formatted = this?.toStringAsFixed(decimalPlaces) ?? '';
+
+    // If the user doesn't want to keep trailing zeros
+    if (!keepTrailingZeros) {
+      formatted = formatted
+          .replaceAll(RegExp(r'0+$'), '')
+          .replaceAll(RegExp(r'\.$'), '');
+    }
+
+    return formatted;
+  }
 }
 
 extension DHUNumExtensions on num {